When uploading files, use "parent_id" instead of "folder_id"
When you upload a file, the documentation states that you should specify
a field called "parent_id" which is the folder id of the folder into
which the file should be uploaded.

The API also accepts "folder_id"; but "parent_id" is the preferred name
for the field. So, this commit changes the Java SDK to use "parent_id"
instead of "folder_id".