Fix bug where token refresh could get stuck in a permanent failure state
If the SDK encountered some server error when attempting to refresh the access token,
the OAuthDataController would get stuck in a permanent failure state. This is appropriate
for the case when we get a 400 back from the server, which indicates our refresh token is
invalid and that the user must log in again. It is not appropriate for other sorts of errors,
such as a lack of connectivity or some other server error thrown by the Box backend. For these
latter types of errors, this commit prevents us from going into a failure state, so the app
can retry.